chore: consolidate registry-related tests
